+++ tests/fake_claude.py
@@ -23,6 +23,8 @@ log_path = os.environ.get("FAKE_CLAUDE_LOG")
 if log_path:
     with open(log_path, "a") as fh:
         fh.write(json.dumps(argv) + "\n")
+        fh.write("ENV BRUH_DENIED_SERVICES="
+                 + os.environ.get("BRUH_DENIED_SERVICES", "") + "\n")
 
 mode = os.environ.get("FAKE_MODE", "ok")